requirements.txt
asgiref==3.8.1
boto3==1.34.119
botocore==1.34.119
Django==5.0.6
django-cors-headers==3.13.0
django-rest-batteries==1.4.1
django-storages==1.14.3
djangorestframework==3.14.0
jmespath==1.0.1
pika==1.3.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
